requirements-arxiv.txt
# arXiv Daily Crawler Requirements

# Core dependencies
arxiv>=2.1.0
feedparser>=6.0.10
requests>=2.31.0
PyYAML>=6.0.1
Jinja2>=3.1.2

# AI providers (optional - uncomment what you need)
openai>=1.12.0
anthropic>=0.18.0

# For advanced features
python-dateutil>=2.8.2
